Initial Setup for React Native w/ Typescript
Using the popular React Native Typescript boiler plate, I created the starting point for the app.
