+++ audio-peeler-web-app/src/mashup/Mashup.js
@@ -1,0 +1,29 @@
+import axios from "axios";
+import './Mashup.css';
+import React, { useState } from 'react';
+import Home from "../home/Home";
+
+function Mashup(){
+
+
+    return (
+        <div className="App">
+            <header className="App-header">
+                <div>
+                    <h4>
+                        Lets
+                    </h4>
+                    <h2>
+                        Mash it up!
+                    </h2>
+                </div>
+
+
+                <div class="output_area">
+                    Welcome to the mashup output area!
+                </div>
+            </header>
+        </div>
+    )};
+
+export default Mashup;
